requirements.txt
kaggle>=1.6.2
python-dotenv==1.0.0